Roll a specialized 1D interpolation and drop the scipy dependency (#1)
The only reason we had scipy as a dependency was usage of interp1 in
a single class. Scipy is relatively heavy for such limited usage.
Additionally, since we should expect to notice appreciable
performance gains from caching, it makes sense to roll our own
interpolant using numpy as a base.