Add WGS84 Ellipsoid and Lat, lon, alt calculations
Earth representation was added to a new file and two classes. One for a
spherical earth and the other for an oblate spheroid (the WGS84
ellipsoid).

New core functionality was added that enables the transformation of
geocentric cartesian coordinates (x, y, z) to geodetic latitude,
longitude, and altitude (height above ellipsoid) or vice versa.